realtime-airline-distributed-database-main/src/backend/pyproject.toml
[tool.poetry]
name = "data_router"
version = "0.1.0"
description = ""
authors = ["Navneet Raju <navneetr@usc.edu>", "Shalini Sai Prasad <saiprasa@usc.edu>", "Nammi Sridhara <knsridha@usc.edu>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
loguru = "^0.7.2"
kazoo = "^2.10.0"
confluent-kafka = "^2.3.0"
flightradarapi = "^1.3.25"
selenium = "^4.18.1"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
